feat(cortex): Implement hierarchical sparse domain routing with tool sidecar and episodic reasoning
- Add lasmoid-cortex architecture specs (architecture_selection.md, design.md, requirements.md)
- Add lasmoid-hardening specs and configuration for production stability verification
- Implement DomainCortexRouter for hierarchical sparse routing across domain columns
- Add cortex.py with domain-steerable token routing and cross-domain affinity logic
- Implement ReasoningController with propose→verify→refine loop and step budgeting
- Add EpisodicMemory integration for long-context working memory clustering
- Implement RelationalCortex for optional multi-track representation refinement
- Add ToolRegistry and EDAToolset for structured tool dispatch and data analysis
- Add reasoning.py and relational.py for reasoning and relational inference modules
- Extend Gate routing to compose cortex mask with expert selection (mask-before-select invariant)
- Add curiosity.py and episodic.py modules for exploration and memory management
- Update MoE, attention, and block implementations to support cortex-gated expert routing
- Add config_1b_2m.json for 1B parameter model with cortex and episodic features
- Implement _harness.py for integration testing and reference parity validation
- Add comprehensive test suite: test_attention_yarn.py, test_cif_correctness.py, test_reference_parity.py, test_sampler.py
- Add GPU benchmark and training scripts (gpu_benchmark.py, gpu_generate.py, gpu_train.py)
- Add training infrastructure: ablation framework (run_ablations.py, ablation_report.md), evaluate.py, report.py
- Update encoding, loss, vision, vq, and sampler modules with cortex-aware processing
- Update prepare_data.py, pretrain.py, long_context_finetune.py, optimizer.py with cortex support
- Add MIGRATIONS.md for version compatibility and feature gating
- Preserve byte-for-byte backward compatibility with use_domain_cortex=False config flag
- Zero synchronous cycles: acyclic information flow between model core, tools, and reasoning controller

+++ inference/vision.py
@@ -167,8 +167,8 @@ class LasmoidVisionEncoder(nn.Module):
             positions_xy = torch.stack([grid_x, grid_y], dim=-1).flatten(0, 1)  # [num_patches, 2]
             positions_xy = positions_xy.unsqueeze(0).expand(B, -1, -1)  # [B, num_patches, 2]
 
-        x_coords = torch.clamp(positions_xy[..., 0], 0, self.pos_emb.size(0) - 1)
-        y_coords = torch.clamp(positions_xy[..., 1], 0, self.pos_emb.size(0) - 1)
+        x_coords = torch.clamp(positions_xy[..., 0], 0, self.pos_emb.size(0) - 1).long()
+        y_coords = torch.clamp(positions_xy[..., 1], 0, self.pos_emb.size(0) - 1).long()
 
         x_emb = self.pos_emb[x_coords, 0, :]
         y_emb = self.pos_emb[y_coords, 1, :]
